+++ front/app/containers/Admin/projects/project/general/components/ProjectCardImageDropzone.tsx
