+++ front/app/containers/Admin/projects/project/general/index.tsx
@@ -13,7 +13,7 @@ import ProjectTypePicker from './components/ProjectTypePicker';
 import TopicInputs from './components/TopicInputs';
 import GeographicAreaInputs from './components/GeographicAreaInputs';
 import HeaderBgUploader from 'components/admin/ProjectableHeaderBgUploader';
-import ProjectCardImageDropzone from './components/ProjectImageDropzone';
+import ProjectCardImageDropzone from './components/ProjectCardImageDropzone';
 import AttachmentsDropzone from './components/AttachmentsDropzone';
 import SubmitWrapper, { ISubmitState } from 'components/admin/SubmitWrapper';
 import {
